Create schema, Database class, and add configs
The database can bre created by running database.py as a script.
It locates where to create the database by reading the value for the
DATABASE_PATH in config.py.
The schema to be used is located by reading the value from the
SCHEMA_PATH constant in config.py, which is the path to schema.sql.