Fix registration page crash and add error handling
Move rerendering of template after if-elif-else statement chacking that
all fields of the form have been filled.
This fixes the page crashing when a field was not filled, as previouslt
no return happened in that senario.

Add error handling for when username is already taken and for
unsuccessful commit the database.
The error handling for unsuccessful commit is not fool proof per
limitations of SQLite3.